// src/app/models/Address.ts
export class Address {
    Property1: Class1[];
    Property2: Class2[];
}

class Class1 {
    busrid: string;
    userrid: string;
    Address1: string;
    Address2: string;
    Country: string;
    State: string;
    City: string;
    Suburb: string;
    zipcode: string;
    Postcode: string;
    Latitude: string;
    Longitude: string;
}
class Class2 {
    Address1: boolean;
    Address2: boolean;
    Country: boolean;
    State: boolean;
    City: boolean;
    Suburb: boolean;
    zipcode: boolean;
    Postcode: boolean;
}